+++ lib/volt/server/message_bus/redis.rb
@@ -1,0 +1,1 @@
+require 'volt/server/message_bus/base_message_bus'